+++ Controller/CampaignController.php
@@ -4,12 +4,12 @@
  * Contains campaign controller.
  */
 
-namespace Itk\CampaignBundle\Controller;
+namespace Os2Display\CampaignBundle\Controller;
 
 use FOS\RestBundle\Controller\Annotations as Rest;
 use FOS\RestBundle\Request\ParamFetcherInterface;
 use Symfony\Component\HttpFoundation\Response;
-use Itk\CampaignBundle\Entity\Campaign;
+use Os2Display\CampaignBundle\Entity\Campaign;
 use Os2Display\CoreBundle\Controller\ApiController;
 use Os2Display\CoreBundle\Exception\DuplicateEntityException;
 use Os2Display\CoreBundle\Exception\HttpDataException;